backend/requirements.txt
Django==4.0
python-dotenv~=0.17.1
psycopg2-binary==2.9.3
asyncpg~=0.23.0
djangorestframework~=3.13.1
redis~=4.3.3
gunicorn~=20.1.0
drf-yasg~=1.20.0
asgiref==3.5.2
pydantic==1.9.2
backoff==2.1.2
django-environ==0.9.0
pyjwt==2.6.0
django-cors-headers==3.14.0
celery~=4.4.2